Added the custom hooks folder Added hooks of some screens and components onto this folder
