+++ src/components/LogOutButton/index.js
@@ -1,5 +1,4 @@
 import React, {useContext} from 'react'
-import { View, Text, Button } from 'react-native'
 import auth from '@react-native-firebase/auth';
 import { LogOut, TextLog } from './styled';
