+++ src/screens/MyFlights.js
@@ -5,35 +5,20 @@ import { FlightDetails, AddButton, LogOutButton } from '../components'
 import { PurpleText } from '../components/Styled/generals'
 import firestore from '@react-native-firebase/firestore';
 import auth from '@react-native-firebase/auth';
+import { MyFlightsHooks } from '../library/CustomHooks/MyFlightsHooks'
 
 export const MyFlights = (props) => {
 
-    const [data, setData] = useState()
-    const [rtData, setRTData] = useState([])
-    const [user, setUser] = useState([])
-
-
-    async function loadData () {
-        try {
-            const flights = await  firestore().collection('Flights').get()
-            setData(flights.docs)
-        } catch (error) {
-            console.log(error);
-        }
-    }
-
-    async function loadRTData() {
-        const subscriber = firestore().collection('Flights').onSnapshot(querySnapshot => {
-            const flights = []
-            querySnapshot.forEach(documentSnapshot => {
-                flights.push({
-                    ...documentSnapshot.data(),
-                    key: documentSnapshot.id
-                })
-            })
-            setRTData(flights)
-        })
-    }
+    const {
+        data,
+        setData,
+        rtData,
+        setRTData,
+        user,
+        setUser,
+        loadData,
+        loadRTData
+    } = MyFlightsHooks();
 
     useEffect(() => {
         loadData()
